--- api/src/db/database.ini
@@ -1,6 +1,0 @@
-[postgresql]
-host=localhost
-port=5432
-dbname=postgres
-user=postgres
-password=postgres
